Changes to be committed: 	modified:   01_SumstatsQC_main_module.sh 	modified:   03_SumstatsQC_sumstats_munge.sh 	modified:   04_SumstatsQC_refsumstats_merge.sh 	modified:   05_SumstatsQC_QCparameters.sh 	modified:   06_SumstatsQC_Postprocessing.sh 	modified:   07_SumstatsQC_extractqcinfo.sh 	modified:   08_SumstatsQC_visualization.sh 	modified:   10_SumstatsQC_archive.sh

+++ 10_SumstatsQC_archive.sh
@@ -41,5 +41,19 @@
     gsutil cp -r $prefix.SumstatsQC.files $bucket
 
     echo "Folder has been archived"
+    
+    while [ 1 ]; do 
+        read -p "Delete folder (y/n)?" CONT1
+        if [ "$CONT1" == "y" ]; then
+            echo "Deleting ...."
+                rm -r $prefix.SumstatsQC.files
+            break
+        elif [ "$CONT1" == "n" ]; then 
+            echo "abort delete folder ...."
+            exit 1
+        else 
+            echo "respond to y or n ? "
+        fi
+    done
 
-    rm -r $prefix.SumstatsQC.files
+
